+++ Api/src/main/java/com/example/springboot/dtos/EstoqueRecordDto.java
@@ -7,7 +7,7 @@ import jakarta.validation.constraints.NotNull;
 
 
 
-public record EstoqueRecordDto(@NotNull Integer quantidade, @NotNull Integer idProduct, Integer idEstoque) {
+public record EstoqueRecordDto(@NotNull Integer quantidade, @NotNull Long idProduct, Integer idEstoque) {
